requirements.txt
andas
numpy
matplotlib
scipy
jupyter
